+++ packages/polaris-viz/CHANGELOG.md
@@ -7,6 +7,10 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## Unreleased -->
 
+## [9.5.1] - 2023-07-05
+
+- No updates. Transitive dependency bump.
+
 ## [9.5.0] - 2023-07-05
 
 ### Fixed